+++ packages/web/example/handlers/users.ts
@@ -1,4 +1,5 @@
 import { Tracer } from "@fewlines/fwl-tracing";
+
 import {
   HandlerPromise,
   HttpStatus,
@@ -7,9 +8,7 @@ import {
   UnmanagedError,
 } from "../../index";
 
-export interface GetUsersByIdParams {
-  id: string;
-}
+export type GetUsersByIdParams = any;
 
 const users = [
   { id: "f55bce0a-a2d3-49a8-8fd6-5a6d2d5b3a18", name: "John Doe" },
@@ -39,7 +38,7 @@ export function getUserById() {
   };
 }
 
-export type CreateUserParams = {};
+export type CreateUserParams = Record<string, unknown>;
 export interface CreateUserBody {
   name: string;
 }