+++ packages/web/example/server.ts
@@ -1,15 +1,15 @@
 import { Logger } from "@fewlines/fwl-logging";
 import { Tracer } from "@fewlines/fwl-tracing";
-import { createApp, loggingMiddleware, Router } from "../index";
 import { Application } from "express";
 
+import { createApp, loggingMiddleware, Router } from "../index";
 import { pingHandler } from "./handlers/ping";
 import * as userHandler from "./handlers/users";
 
 export function start(tracer: Tracer, logger: Logger): Application {
   const router = new Router(tracer, logger);
 
-  router.get<{}>("/ping", pingHandler());
+  router.get<Record<string, unknown>>("/ping", pingHandler());
   router.get<userHandler.GetUsersByIdParams>(
     "/users/:id",
     userHandler.getUserById(),